- replaced the small member icon for online/offline status with some bigger ones and also increased the font size of the links. - started moving includes/triggers.inc.php to regular event files - comments on any item are now handled from the same processor for all items instead of 1 processor per item. - started implementing latest_blog_comments widget

+++ trunk/admin/processors/photo_delete.php
@@ -28,16 +28,25 @@ if (!($res=@mysql_query($query))) {trigger_error(mysql_error(),E_USER_ERROR);}
 if (mysql_num_rows($res)) {
 	$input=mysql_fetch_assoc($res);
 
-	on_before_delete_photo(array($photo_id));
-
 	$query="DELETE FROM `{$dbtable_prefix}user_photos` WHERE `photo_id`=$photo_id";
+	if (is_file(_BASEPATH_.'/events/processors/photo_delete.php')) {
+		include_once _BASEPATH_.'/events/processors/photo_delete.php';
+		if (isset($_on_before_delete)) {
+			$GLOBALS['photo_ids']=array($photo_id);
+			for ($i=0;isset($_on_before_delete[$i]);++$i) {
+				call_user_func($_on_before_delete[$i]);
+			}
+		}
+	}
 	if (!($res=@mysql_query($query))) {trigger_error(mysql_error(),E_USER_ERROR);}
 
-	require_once '../../includes/classes/fileop.class.php';
-	$fileop=new fileop();
-	$fileop->delete(_PHOTOPATH_.'/t1/'.$input['photo']);
-	$fileop->delete(_PHOTOPATH_.'/t2/'.$input['photo']);
-	$fileop->delete(_PHOTOPATH_.'/'.$input['photo']);
+	if (!empty($input['photo'])) {
+		require_once '../../includes/classes/fileop.class.php';
+		$fileop=new fileop();
+		$fileop->delete(_PHOTOPATH_.'/t1/'.$input['photo']);
+		$fileop->delete(_PHOTOPATH_.'/t2/'.$input['photo']);
+		$fileop->delete(_PHOTOPATH_.'/'.$input['photo']);
+	}
 
 	$query="DELETE FROM `{$dbtable_prefix}photo_comments` WHERE `fk_parent_id`=$photo_id";
 	if (!($res=@mysql_query($query))) {trigger_error(mysql_error(),E_USER_ERROR);}
